move class hash to utils
